GCP/requirements.txt
tensorflow==2.18.0
pillow==11.1.0
google-cloud-storage==2.19.0